test: Merge codecov reports (#124)
Resolves: #111
