+++ .github/workflows/build.yaml
@@ -35,11 +35,6 @@ jobs:
         run: GOPATH="$(go env GOPATH)" ./hack/verify-codegen.sh
       - name: Run unit tests
         run: make unit-tests
-      - name: Upload code coverage
-        uses: codecov/codecov-action@v1
-        with:
-          token: ${{ secrets.CODECOV_TOKEN }}
-          file: ./coverage.txt
       - name: Setup Kubernetes cluster (KIND)
         uses: engineerd/setup-kind@v0.4.0
         with:
@@ -54,6 +49,10 @@ jobs:
         env:
           KUBECONFIG: /home/runner/.kube/config
           STARBOARD_CLI_LOG_LEVEL: "0"
+      - name: Upload code coverage
+        uses: codecov/codecov-action@v1
+        with:
+          files: ./coverage.txt,./itest/coverage.txt
       - name: Release snapshot
         uses: goreleaser/goreleaser-action@v2
         with:
