+++ .gitignore
@@ -4,6 +4,8 @@ bin/
 dist/
 
 *.html
+
 coverage.txt
+itest/coverage.txt
 
 vendor
