+++ pkg/cmd/root.go
@@ -1,11 +1,18 @@
 package cmd
 
 import (
+	"flag"
+	"io"
+	"path/filepath"
+	"strings"
+
+	"github.com/spf13/pflag"
+
 	"github.com/spf13/cobra"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 )
 
-func NewRootCmd(executable string, version VersionInfo) *cobra.Command {
+func NewRootCmd(version VersionInfo, args []string, outWriter io.Writer, errWriter io.Writer) *cobra.Command {
 	var cf *genericclioptions.ConfigFlags
 
 	rootCmd := &cobra.Command{
@@ -16,6 +23,8 @@ func NewRootCmd(executable string, version VersionInfo) *cobra.Command {
 	}
 
 	cf = genericclioptions.NewConfigFlags(true)
+
+	executable := executable(args)
 
 	rootCmd.AddCommand(NewVersionCmd(version))
 	rootCmd.AddCommand(NewInitCmd(cf))
@@ -28,5 +37,36 @@ func NewRootCmd(executable string, version VersionInfo) *cobra.Command {
 
 	SetGlobalFlags(cf, rootCmd)
 
+	rootCmd.SetArgs(args[1:])
+	rootCmd.SetOut(outWriter)
+	rootCmd.SetErr(errWriter)
+
 	return rootCmd
+}
+
+func executable(args []string) string {
+	if strings.HasPrefix(filepath.Base(args[0]), "kubectl-") {
+		return "kubectl starboard"
+	}
+	return "starboard"
+}
+
+// Run is the entry point of the Starboard CLI. It runs the specified
+// command based on the specified args.
+func Run(version VersionInfo, args []string, outWriter io.Writer, errWriter io.Writer) error {
+
+	initFlags()
+
+	return NewRootCmd(version, args, outWriter, errWriter).Execute()
+}
+
+func initFlags() {
+	pflag.CommandLine.AddGoFlagSet(flag.CommandLine)
+
+	// Hide all klog flags except for -v
+	flag.CommandLine.VisitAll(func(f *flag.Flag) {
+		if f.Name != "v" {
+			pflag.Lookup(f.Name).Hidden = true
+		}
+	})
 }
